fix(selfupdate): Windows 파일 ID 검증을 안정화한다
Constraint: 범위 밖 회귀 금지

🐙 Autopus <noreply@autopus.co>
